assignment_code/src/requirements.txt
Django==1.11.29
glfw==2.8.0
image==1.5.32
pillow==11.0.0
PyOpenGL==3.1.7
pytz==2020.1
six==1.15.0
